refactor: 공통된 인터페이스 types로 정리
