+++ hooks/src/lib/useExpiryYear.ts
@@ -1,4 +1,5 @@
-import useInput, { ValidationResult } from './useInput';
+import { CustomValidator } from './types';
+import useInput from './useInput';
 import { useEffect } from 'react';
 
 const validateInputType = (value: string) => {
@@ -27,10 +28,8 @@ const validateFieldRules = (value: string) => {
   return { isValid: true, errorMessage: '' };
 };
 
-export interface ExpiryYearOptions {
+export interface ExpiryYearOptions extends CustomValidator {
   isAutoFocus?: boolean;
-  customValidateInputType?: (value: string) => ValidationResult;
-  customValidateFieldRules?: (value: string) => ValidationResult;
 }
 
 const useExpiryYear = (initialValue: string, options?: ExpiryYearOptions) => {
