+++ hooks/src/lib/useInput.ts
@@ -1,14 +1,5 @@
 import { useState } from 'react';
-
-export interface ValidationResult {
-  isValid: boolean;
-  errorMessage: string;
-}
-
-interface ValidatorProps {
-  validateInputType: (value: string) => ValidationResult;
-  validateFieldRules: (value: string) => ValidationResult;
-}
+import { ValidationResult, ValidatorProps } from './types';
 
 const useInput = (initialValue: string, validator: ValidatorProps) => {
   const { validateInputType, validateFieldRules } = validator;
